requirements.txt
purpleflea>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
langchain>=0.3.0
langchain-openai>=0.2.0
crewai>=0.80.0
pyautogen>=0.4.0
requests>=2.31.0
